+++ src/constants.ts
@@ -1,7 +1,9 @@
 // Action events
 export const EVENT_PULL_REQUEST = "pull_request";
 export const EVENT_PUSH = "push";
+export const EVENT_REPOSITORY_DISPATCH = "repository_dispatch";
 export const EVENT_SCHEDULE = "schedule";
+export const EVENT_WORKFLOW_DISPATCH = "workflow_dispatch";
 
 // Action configuration
 export const INPUT_NAME_BRANCH = "branch";